More updates, move into src folder
